+++ src/cpp/parquet_writer_helpers.cpp
@@ -1,4 +1,4 @@
-#include "parquet_helpers.h"
+#include "parquet_writer_helpers.h"
 
 #include "parquet_writer_exceptions.h"
