+++ src/cpp/parquet_writer_helpers.h
